Move search scheduler to search.rs
